Fix performance issue with Effect.getselected()
Combine and rewrite the two Effect() methods to build dicts with id,
node of selected objects and use count of ids in the whole document to
avoid the slow performance of the XPath expression used for every id
passed via arguments to find the ElementTree node corresponding to the
selected object. Overload Effect.affect() to call the new combined
method instead of the two original ones.

Ideally, a similar solution would be applied to inkex.py (upstream), or
at least to inkex_local.py (preferred by inx-pathops over inkex.py if
available).